+++ ReleaseNotes.md
@@ -3,6 +3,7 @@
 ## v1.15.0
 * **NEW FEATURE:** Added method `QcsSessionInfo.GetJObject()` to simplify interaction with the .NET SDK.
 * **NEW FEATURE:** Added additional endpoints accepting `HttpContent` as argument.
+* **EXAMPLE:** Added example `QcsUploadDataFile`.
 * **ENVIRONMENTAL:** Added target framework .NET 7.0.
 * **ENVIRONMENTAL:** Added nuget dependency on `System.IdentityModel.Tokens.Jwt` as this is necessary for JWT creation.